move(Profanity, KitchenRiders): interface domain계층으로 이동
