+++ src/main/java/kitchenpos/common/domain/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.infra;
+package kitchenpos.common.domain;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
