+++ src/main/java/kitchenpos/order/deliveryorder/domain/KitchenridersClient.java
@@ -1,8 +1,9 @@
-package kitchenpos.order.deliveryorder.infra;
+package kitchenpos.order.deliveryorder.domain;
 
 import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface KitchenridersClient {
+
     void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
 }
