+++ src/main/java/kitchenpos/order/deliveryorder/infra/DefaultKitchenridersClient.java
@@ -1,12 +1,13 @@
 package kitchenpos.order.deliveryorder.infra;
 
-import org.springframework.stereotype.Component;
-
 import java.math.BigDecimal;
 import java.util.UUID;
+import kitchenpos.order.deliveryorder.domain.KitchenridersClient;
+import org.springframework.stereotype.Component;
 
 @Component
 public class DefaultKitchenridersClient implements KitchenridersClient {
+
     @Override
     public void requestDelivery(final UUID orderId, final BigDecimal amount, final String deliveryAddress) {
     }
